https://issues.apache.org/jira/browse/EXTSCRIPT-18
Ongoing works on the initial scan infrastructure.

The initial scripting annotation scan has to be performed
after myfaces has started up and before the first page is accessed
which means we use the jsf2 facilities of application events
to target the right lifecycle point.

The idea is to wait for postConstructApplication and then trigger
the initial source scan.
By then the scripting subsystem weavers are initialized
and we can trigger the initial scan on the weavers which then
can add the missing source artefacts to the runtime config holding
all the faces-config meta data!

It still is undecided yet whether the weavers themselves
can register their scanners dynamically or whether
the startup listener registers them in the weavers.
(I have to sleep over that one)

git-svn-id: https://svn.apache.org/repos/asf/myfaces/extensions/scripting/trunk@816646 13f79535-47bb-0310-9956-ffa450edef68