https://issues.apache.org/jira/browse/EXTSCRIPT-74
We now have replaced the old list based system with a first rough version of a registry / strategy based one which works
on filter lists. While we were clearing this up we also improved the scan time in the whitelist filters by
moving from a string based compare system to one based on index maps.
The exection time was reduced by 50% just by this measure (being run on a loop which looped over a scan on 1000 classes
for quick performance testing)

git-svn-id: https://svn.apache.org/repos/asf/myfaces/extensions/scripting/trunk@916619 13f79535-47bb-0310-9956-ffa450edef68